requirements.txt
requests >= 2.30.0
pyyaml >= 6.0
plexapi >= 4.13.4